+++ src/main/java/com/ironhacker/JavaPottersHomework3/repository/ContactRepository.java
@@ -1,10 +1,12 @@
 package com.ironhacker.JavaPottersHomework3.repository;
 
 import com.ironhacker.JavaPottersHomework3.models.Contact;
+import com.ironhacker.JavaPottersHomework3.models.Lead;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContactRepository extends JpaRepository<Contact, Long> {
+public interface ContactRepository extends JpaRepository<Contact, Integer> {
+
 }
